docs and provide options to client
