+++ sdk/nextjs/package.json
@@ -10,7 +10,7 @@
         "server-only": "^0.0.1"
     },
     "peerDependencies": {
-        "next": "^14.0.5",
+        "next": "^14.0.0",
         "react": "^18.2.0"
     },
     "exports": {
